[refactor] User 수정/삭제 FE 부분 리팩토링 #25
