+++ src/main/resources/templates/customer/main.html
@@ -142,8 +142,15 @@
                     Authorization: 'Bearer ' + token
                 }
             }).then(res => {
-                sessionStorage.setItem("accessToken", res.data.data.accessToken);
-                getUserAddress();
+                if(res.data.message === '토큰 정상'){
+                    getUserAddress();
+                } else if(res.data.message === '성공'){
+                    sessionStorage.setItem("accessToken", res.data.data.accessToken);
+                    getUserAddress();
+                } else {
+                    alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
+                    location.href = '/';
+                }
             }).catch(error => {
                 alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
                 location.href = '/';
