+++ src/main/resources/templates/customer/my-info.html
@@ -78,8 +78,15 @@
                     Authorization: 'Bearer ' + token
                 }
             }).then(res => {
-                sessionStorage.setItem("accessToken", res.data.accessToken);
-                getUserInfo();
+                if(res.data.message === '토큰 정상'){
+                    getUserInfo();
+                } else if(res.data.message === '성공'){
+                    sessionStorage.setItem("accessToken", res.data.data.accessToken);
+                    getUserInfo();
+                } else {
+                    alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
+                    location.href = '/';
+                }
             }).catch(error => {
                 alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
                 location.href = '/';
@@ -92,10 +99,10 @@
                     Authorization: 'Bearer ' + token
                 }
             }).then(function(response) {
-                document.getElementById('email').innerHTML = response.data.email;
-                document.getElementById('name').innerHTML = response.data.name
-                document.getElementById('phone').innerHTML = response.data.phone
-                document.getElementById('address').innerHTML = response.data.address
+                document.getElementById('email').innerHTML = response.data.data.email;
+                document.getElementById('name').innerHTML = response.data.data.name
+                document.getElementById('phone').innerHTML = response.data.data.phone
+                document.getElementById('address').innerHTML = response.data.data.baseAddress
             }).catch(function(error) {
                 console.error(error);
             });
