+++ src/main/resources/templates/customer/my-page.html
@@ -77,8 +77,15 @@
           Authorization: 'Bearer ' + token
         }
       }).then(res => {
-        sessionStorage.setItem("accessToken", res.data.accessToken);
-        getUserRole();
+        if(res.data.message === '토큰 정상'){
+          getUserRole();
+        } else if(res.data.message === '성공'){
+          sessionStorage.setItem("accessToken", res.data.data.accessToken);
+          getUserRole();
+        } else {
+          alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
+          location.href = '/';
+        }
       }).catch(error => {
         alert("로그인이 만료되었습니다. 다시 로그인해주세요.");
         location.href = '/';
@@ -104,8 +111,8 @@
           Authorization: 'Bearer ' + token
         }
       }).then(() => {
-        alert("로그아웃 되었습니다.");
         sessionStorage.removeItem("accessToken");
+        alert("로그아웃 되었습니다.");
         location.href = '/';
       }).catch((error) => {
         console.error(error);
